+++ test/test_helper.rb
@@ -10,6 +10,7 @@ end
 
 require File.expand_path "#{File.dirname __FILE__}/../../../test/test_helper"
 require File.expand_path "#{File.dirname __FILE__}/global_test_helper"
+require File.expand_path "#{File.dirname __FILE__}/crud_controller_base"
 
 module Additionals
   module TestHelper