+++ app/helpers/additionals_chartjs_helper.rb
@@ -6,7 +6,7 @@ module AdditionalsChartjsHelper
   end
 
   def select_options_for_chartjs_colorscheme(selected)
-    data = YAML.safe_load(ERB.new(IO.read(Rails.root.join('plugins/additionals/config/colorschemes.yml'))).result) || {}
+    data = YAML.safe_load(ERB.new(IO.read(File.join(Additionals.plugin_dir, 'config', 'colorschemes.yml'))).result) || {}
     grouped_options_for_select(data, selected)
   end
 end