+++ app/jobs/additionals_remove_unused_tag_job.rb
@@ -1,5 +1,0 @@
-class AdditionalsRemoveUnusedTagJob < AdditionalsJob
-  def perform
-    AdditionalsTag.remove_unused_tags
-  end
-end