+++ app/models/additionals_query.rb
@@ -150,7 +150,7 @@ module AdditionalsQuery
   end
 
   def sql_for_tags_field(field, _operator, value)
-    AdditionalsTag.sql_for_tags_field(queried_class, operator_for(field), value)
+    AdditionalTags.sql_for_tags_field(queried_class, operator_for(field), value)
   end
 
   def sql_for_is_private_field(_field, operator, value)