+++ app/overrides/account/register.rb
@@ -1,5 +1,0 @@
-Deface::Override.new virtual_path: 'account/register',
-                     name: 'add-invisble-captcha',
-                     insert_top: 'div.box',
-                     original: 'a9c303821376a8d83cba32654629d71cc3926a1d',
-                     partial: 'account/invisible_captcha'