+++ lib/additionals.rb
@@ -1,3 +1,5 @@
+require 'additionals/version'
+
 module Additionals
   MAX_CUSTOM_MENU_ITEMS = 5
   SELECT2_INIT_ENTRIES = 20
@@ -5,17 +7,16 @@ module Additionals
   GOTO_LIST = " \xc2\xbb".freeze
   LIST_SEPARATOR = "#{GOTO_LIST} ".freeze
 
-  RenderAsync.configuration.jquery = true
-
   class << self
     def setup
+      RenderAsync.configuration.jquery = true
+
       incompatible_plugins %w[redmine_issue_control_panel
                               redmine_editauthor
                               redmine_changeauthor
                               redmine_auto_watch]
 
-      patch %w[AccountController
-               ApplicationController
+      patch %w[ApplicationController
                AutoCompletesController
                Issue
                IssuePriority
@@ -60,10 +61,7 @@ module Additionals
       require_dependency 'additionals/hooks'
 
       # Macros
-      load_macros %w[cryptocompare date fa gist gmap google_docs group_users iframe
-                     issue redmine_issue redmine_wiki
-                     last_updated_at last_updated_by meteoblue member new_issue project
-                     recently_updated reddit slideshare tradingview twitter user vimeo youtube asciinema]
+      load_macros
     end
 
     # support with default setting as fall back
@@ -117,10 +115,19 @@ module Additionals
       end
     end
 
-    def load_macros(macros = [], plugin_id = 'additionals')
-      macro_dir = Rails.root.join("plugins/#{plugin_id}/lib/#{plugin_id}/wiki_macros")
-      macros.each do |macro|
-        require_dependency "#{macro_dir}/#{macro.underscore}_macro"
+    def load_macros(plugin_id = 'additionals')
+      Dir[File.join(plugin_dir(plugin_id),
+                    'lib',
+                    plugin_id,
+                    'wiki_macros',
+                    '**/*_macro.rb')].sort.each { |f| require f }
+    end
+
+    def plugin_dir(plugin_id = 'additionals')
+      if Gem.loaded_specs[plugin_id].nil?
+        File.join(Redmine::Plugin.directory, plugin_id)
+      else
+        Gem.loaded_specs[plugin_id].full_gem_path
       end
     end
 
@@ -128,7 +135,7 @@ module Additionals
       cached_settings_name = "@load_settings_#{plugin_id}"
       cached_settings = instance_variable_get cached_settings_name
       if cached_settings.nil?
-        data = YAML.safe_load(ERB.new(IO.read(Rails.root.join("plugins/#{plugin_id}/config/settings.yml"))).result) || {}
+        data = YAML.safe_load(ERB.new(IO.read(File.join(plugin_dir(plugin_id), '/config/settings.yml'))).result) || {}
         instance_variable_set cached_settings_name, data.symbolize_keys
       else
         cached_settings
@@ -150,6 +157,29 @@ module Additionals
 
     def settings
       Setting[:plugin_additionals]
+    end
+  end
+
+  # Run the classic redmine plugin initializer after rails boot
+  class Plugin < ::Rails::Engine
+    require 'deface'
+    require 'emoji'
+    require 'render_async'
+    require 'rss'
+
+    config.after_initialize do
+      # engine_name could be used (additionals_plugin), but can
+      # create some side effencts
+      plugin_id = 'additionals'
+
+      # if plugin is already in plugins directory, use this and leave here
+      next if Redmine::Plugin.installed? plugin_id
+
+      # gem is used as redmine plugin
+      require File.expand_path '../init', __dir__
+      AdditionalTags.setup
+      Additionals::Gemify.install_assets plugin_id
+      Additionals::Gemify.create_plugin_hint plugin_id
     end
   end
 end