+++ test/unit/additionals_test.rb
@@ -50,10 +50,9 @@ class AdditionalsTest < Additionals::TestCase
   end
 
   def test_load_macros
-    assert_equal ['fa'], Additionals.load_macros(['fa'])
+    macros = Additionals.load_macros
 
-    assert_raises LoadError do
-      Additionals.load_macros(%w[fa invalid])
-    end
+    assert macros.count.positive?
+    assert(macros.detect { |macro| macro.include? 'fa_macro' })
   end
 end