+++ config/routes.rb
@@ -1,6 +1,4 @@
 Rails.application.routes.draw do
-  # Don't create routes for repositories resources with only: []
-  # do not override Redmine's routes.
   resources :issues, only: [] do
     resource 'assign_to_me', only: %i[update], controller: 'additionals_assign_to_me'
     resource 'change_status', only: %i[update], controller: 'additionals_change_status'