+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 - Do not show "Assign to me" if assigned_to is disabled for tracker
 - FontAwesome 5.15.1 support
 - D3 6.2.0 support
+- Mermaid 8.8.2 support
 
 3.0.0
 +++++