+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Changelog
 - Mermaid 8.8.2 support
 - add current_user as special login name for user macro (which shows current login user)
 - add text parameter to user macro (which disable link to user)
+- add asynchronous text block
 
 3.0.0
 +++++