+++ CHANGELOG.rst
@@ -7,12 +7,13 @@ Changelog
 - Do not show "Assign to me" if assigned_to is disabled for tracker
 - FontAwesome 5.15.1 support
 - D3 6.2.0 support
-- Mermaid 8.8.2 support
+- Mermaid 8.8.3 support
 - add current_user as special login name for user macro (which shows current login user)
 - add text parameter to user macro (which disable link to user)
 - add asynchronous text block
 - gemify plugin to use it with Gemfile.local or other plugins
 - remove spam protection functionality
+- Chart.js 2.9.4 support
 
 3.0.0
 +++++